24.11.02 서버 재배포: script/start.sh 조건문 추가

+++ scripts/start.sh
@@ -14,8 +14,9 @@ ERROR_LOG="$PROJECT_ROOT/error.log"
 DEPLOY_LOG="$PROJECT_ROOT/deploy.log"
 
 TIME_NOW=$(date +%c)
-
-kill "$(lsof -t -i:11000)"
+if [ "$(lsof -t -i:11000)" ]; then
+  kill "$(lsof -t -i:11000)"
+fi
 
 chmod +x $JAR_FILE
